requirements.txt
Jinja2>=3.0
WeasyPrint>=56.0
websocket-client>=1.6.2
playwright>=1.40.0
